Refactored ParallelDataset in its own file.
