+++ ml/data/data.go
@@ -21,6 +21,7 @@ import (
 	"crypto/sha256"
 	"encoding/hex"
 	"fmt"
+	"github.com/pkg/errors"
 	"github.com/schollz/progressbar/v3"
 	"io"
 	"log"
@@ -29,13 +30,7 @@ import (
 	"os/exec"
 	"os/user"
 	"path"
-	"runtime"
 	"strings"
-	"sync"
-
-	"github.com/gomlx/gomlx/ml/train"
-	"github.com/gomlx/gomlx/types/tensor"
-	"github.com/pkg/errors"
 )
 
 // FileExists returns true if file or directory exists.
@@ -300,205 +295,4 @@ func Unzip(zipFile, zipBaseDir string) error {
 		return errors.Wrapf(err, "failed to run %q", cmd)
 	}
 	return nil
-}
-
-// ParallelDataset is a wrapper around a `train.Dataset` that parallelize calls to Yield.
-type ParallelDataset struct {
-	Dataset train.Dataset
-
-	// Parallelism is the number of goroutines started generating examples.
-	Parallelism int
-
-	// BufferSize is the size of the cache of pre-generated batches.
-	BufferSize int
-
-	// impl is the actual implementation.
-	impl *parallelDatasetImpl
-
-	// keepAlive is used only to keep ParallelDataset alive in the middle of long calls.
-	keepAlive int64
-}
-
-type yieldUnit struct {
-	spec   any
-	inputs []tensor.Tensor
-	labels []tensor.Tensor
-}
-
-// parallelDatasetImpl separates the implementation of ParallelDataset. It's important
-// that it doesn't point back to the original ParallelDataset, so garbage collecting
-// will also stop the goroutines.
-type parallelDatasetImpl struct {
-	config ParallelDataset // A copy of ht
-
-	err   error
-	muErr sync.Mutex
-
-	cache                                 chan yieldUnit
-	epochFinished, stopEpoch, stopDataset chan struct{}
-}
-
-// NewParallelDataset starts building a ParallelDataset that can be used to parallelize any
-// train.Dataset, as long as the underlying dataset ds is thread-safe.
-//
-// parallelism is the number of goroutines to start, each calling `ds.Yield()` in parallel
-// to accelerate the generation of batches. Set to 0, and it will use the number of cores in
-// the system plus 1.
-//
-// bufferSize is the space reserved to keep pre-generated batches. This is on top of the batches
-// being generated in each goroutine.
-//
-// To avoid leaking goroutines, call ParallelDataset.Cancel when exiting.
-func NewParallelDataset(ds train.Dataset, parallelism, bufferSize int) *ParallelDataset {
-	if parallelism == 0 {
-		parallelism = runtime.NumCPU() + 1
-	}
-	impl := &parallelDatasetImpl{
-		cache:       make(chan yieldUnit, bufferSize),
-		stopDataset: make(chan struct{}),
-	}
-	pd := &ParallelDataset{
-		Dataset:     ds,
-		Parallelism: parallelism,
-		BufferSize:  bufferSize,
-	}
-	impl.config = *pd
-	pd.impl = impl
-	// If the ParallelDataset is garbage collected, stop all parallel goroutines.
-	runtime.SetFinalizer(pd, func(pd *ParallelDataset) {
-		if pd.impl != nil {
-			close(pd.impl.stopDataset)
-			pd.impl = nil
-		}
-	})
-
-	// Start goroutines
-	impl.startGoRoutines()
-	return pd
-}
-
-func (impl *parallelDatasetImpl) startGoRoutines() {
-	impl.epochFinished = make(chan struct{})
-	impl.stopEpoch = make(chan struct{})
-	var wg sync.WaitGroup
-	for ii := 0; ii < impl.config.Parallelism; ii++ {
-		// Start all goroutines.
-		wg.Add(1)
-		go func(impl *parallelDatasetImpl) {
-			defer wg.Done()
-			for {
-				select {
-				case <-impl.stopEpoch:
-					return
-				case <-impl.stopDataset:
-					return
-				default:
-					// Move forward and generate the next batch.
-				}
-				var unit yieldUnit
-				var err error
-				unit.spec, unit.inputs, unit.labels, err = impl.config.Dataset.Yield()
-				if err == io.EOF {
-					return
-				}
-				if err != nil {
-					log.Printf("Error: %+v", err)
-					// Fatal error, stop everything.
-					impl.muErr.Lock()
-					if impl.err != nil {
-						impl.err = err
-					}
-					close(impl.stopEpoch)
-					close(impl.stopDataset)
-					impl.muErr.Unlock()
-					return
-				}
-				select {
-				case <-impl.stopEpoch:
-					return
-				case <-impl.stopDataset:
-					return
-				case impl.cache <- unit:
-					// Batch generated and cached, move to next.
-					continue
-				}
-			}
-		}(impl)
-	}
-
-	// Start controller job.
-	go func() {
-		wg.Wait()
-		impl.muErr.Lock()
-		defer impl.muErr.Unlock()
-		select {
-		case <-impl.stopDataset:
-			return
-		default:
-			//
-		}
-		close(impl.epochFinished)
-	}()
-}
-
-// Name implements train.Dataset.
-func (pd *ParallelDataset) Name() string {
-	return pd.Dataset.Name()
-}
-
-// Reset implements train.Dataset.
-func (pd *ParallelDataset) Reset() {
-	impl := pd.impl
-	impl.muErr.Lock()
-	close(impl.stopEpoch) // Indicate to goroutines to stop generating batches.
-	impl.muErr.Unlock()
-	select {
-	case <-impl.stopDataset:
-		// Return immediately, do nothing.
-		return
-	case <-impl.cache:
-		// Discard remaining entries in cache.
-	case <-impl.epochFinished:
-		// All finished, we can move on.
-	}
-
-	// Reset underlying dataset and start again.
-	impl.config.Dataset.Reset()
-	impl.startGoRoutines()
-
-	// This no-op prevents `pd` from being garbage collected and the goroutines killed in the middle
-	// of the Reset operation. Leave this at the end.
-	pd.keepAlive++
-}
-
-// Yield implements train.Dataset.
-func (pd *ParallelDataset) Yield() (spec any, inputs []tensor.Tensor, labels []tensor.Tensor, err error) {
-	impl := pd.impl
-	var unit yieldUnit
-	select {
-	case <-impl.stopDataset:
-		// An error occurred, dataset is closed.
-		impl.muErr.Lock()
-		err = impl.err
-		impl.muErr.Unlock()
-		return
-	case unit = <-impl.cache:
-		// We got a new batch
-	case <-impl.epochFinished:
-		// No more records being produced (until Reset() is called), but we still need to exhaust the cache.
-		select {
-		case unit = <-impl.cache:
-			// We got a new batch, simply continue.
-		default:
-			// Generation exhausted, and no more records in cache.
-			err = io.EOF
-			return
-		}
-	}
-	spec, inputs, labels = unit.spec, unit.inputs, unit.labels
-
-	// This no-op prevents `pd` from being garbage collected and the goroutines killed in the middle
-	// of the Yield operation. Leave this at the end.
-	pd.keepAlive++
-	return
 }
